+++ lib/api-request/index.spec.js
@@ -5,7 +5,7 @@ var sinon = require('sinon');
 var Request = require('./');
 var SuiteRequest = require('escher-suiteapi-js');
 
-describe('InternalApiRequest', function() {
+describe('ApiRequest', function() {
 
   describe('#get', function() {
